Create generate sun file
Relates #27
